+++ cedar-policy-core/src/ast/entity.rs
@@ -290,9 +290,15 @@ pub struct Entity {
     /// `RestrictedExpr`s, for mostly historical reasons.
     attrs: BTreeMap<SmolStr, PartialValueSerializedAsExpr>,
 
-    /// Set of ancestors of this `Entity` (i.e., all direct and transitive
-    /// parents), as UIDs
-    ancestors: HashSet<EntityUID>,
+    /// Set of indirect ancestors of this `Entity` as UIDs
+    indirect_ancestors: HashSet<EntityUID>,
+
+    /// Set of direct ancestors (i.e., parents) as UIDs
+    ///
+    /// indirect_ancestors and parents should be disjoint
+    /// even if a parent is also an indirect parent through
+    /// a different parent
+    parents: HashSet<EntityUID>,
 
     /// Tags on this entity (RFC 82)
     ///
@@ -318,7 +324,8 @@ impl Entity {
     pub fn new(
         uid: EntityUID,
         attrs: impl IntoIterator<Item = (SmolStr, RestrictedExpr)>,
-        ancestors: HashSet<EntityUID>,
+        indirect_ancestors: HashSet<EntityUID>,
+        parents: HashSet<EntityUID>,
         tags: impl IntoIterator<Item = (SmolStr, RestrictedExpr)>,
         extensions: &Extensions<'_>,
     ) -> Result<Self, EntityAttrEvaluationError> {
@@ -345,7 +352,8 @@ impl Entity {
         Ok(Entity {
             uid,
             attrs: evaluated_attrs,
-            ancestors,
+            indirect_ancestors,
+            parents,
             tags: evaluated_tags,
         })
     }
@@ -361,13 +369,15 @@ impl Entity {
     pub fn new_with_attr_partial_value(
         uid: EntityUID,
         attrs: impl IntoIterator<Item = (SmolStr, PartialValue)>,
-        ancestors: HashSet<EntityUID>,
+        indirect_ancestors: HashSet<EntityUID>,
+        parents: HashSet<EntityUID>,
         tags: impl IntoIterator<Item = (SmolStr, PartialValue)>,
     ) -> Self {
         Self::new_with_attr_partial_value_serialized_as_expr(
             uid,
             attrs.into_iter().map(|(k, v)| (k, v.into())).collect(),
-            ancestors,
+            indirect_ancestors,
+            parents,
             tags.into_iter().map(|(k, v)| (k, v.into())).collect(),
         )
     }
@@ -379,13 +389,15 @@ impl Entity {
     pub fn new_with_attr_partial_value_serialized_as_expr(
         uid: EntityUID,
         attrs: BTreeMap<SmolStr, PartialValueSerializedAsExpr>,
-        ancestors: HashSet<EntityUID>,
+        indirect_ancestors: HashSet<EntityUID>,
+        parents: HashSet<EntityUID>,
         tags: BTreeMap<SmolStr, PartialValueSerializedAsExpr>,
     ) -> Self {
         Entity {
             uid,
             attrs,
-            ancestors,
+            indirect_ancestors,
+            parents,
             tags,
         }
     }
@@ -405,14 +417,34 @@ impl Entity {
         self.tags.get(tag).map(|v| v.as_ref())
     }
 
-    /// Is this `Entity` a descendant of `e` in the entity hierarchy?
+    /// Is this `Entity` a (direct or indirect) descendant of `e` in the entity hierarchy?
     pub fn is_descendant_of(&self, e: &EntityUID) -> bool {
-        self.ancestors.contains(e)
+        self.parents.contains(e) || self.indirect_ancestors.contains(e)
     }
 
-    /// Iterate over this entity's ancestors
+    /// Is this `Entity` a an indirect descendant of `e` in the entity hierarchy?
+    pub fn is_indirect_descendant_of(&self, e: &EntityUID) -> bool {
+        self.indirect_ancestors.contains(e)
+    }
+
+    /// Is this `Entity` a direct decendant (child) of `e` in the entity hierarchy?
+    pub fn is_child_of(&self, e: &EntityUID) -> bool {
+        self.parents.contains(e)
+    }
+
+    /// Iterate over this entity's (direct or indirect) ancestors
     pub fn ancestors(&self) -> impl Iterator<Item = &EntityUID> {
-        self.ancestors.iter()
+        self.parents.iter().chain(self.indirect_ancestors.iter())
+    }
+
+    /// Iterate over this entity's indirect ancestors
+    pub fn indirect_ancestors(&self) -> impl Iterator<Item = &EntityUID> {
+        self.indirect_ancestors.iter()
+    }
+
+    /// Iterate over this entity's direct ancestors (parents)
+    pub fn parents(&self) -> impl Iterator<Item = &EntityUID> {
+        self.parents.iter()
     }
 
     /// Get the number of attributes on this entity
@@ -450,16 +482,22 @@ impl Entity {
         Self {
             uid,
             attrs: BTreeMap::new(),
-            ancestors: HashSet::new(),
+            indirect_ancestors: HashSet::new(),
+            parents: HashSet::new(),
             tags: BTreeMap::new(),
         }
     }
 
     /// Test if two `Entity` objects are deep/structurally equal.
     /// That is, not only do they have the same UID, but also the same
-    /// attributes, attribute values, and ancestors.
+    /// attributes, attribute values, and ancestors/parents.
+    ///
+    /// Does not test that they have the same _direct_ parents, only that they have the same overall ancestor set.
     pub(crate) fn deep_eq(&self, other: &Self) -> bool {
-        self.uid == other.uid && self.attrs == other.attrs && self.ancestors == other.ancestors
+        self.uid == other.uid
+            && self.attrs == other.attrs
+            && (self.ancestors().collect::<HashSet<_>>())
+                == (other.ancestors().collect::<HashSet<_>>())
     }
 
     /// Set the UID to the given value.
@@ -497,30 +535,68 @@ impl Entity {
         Ok(())
     }
 
-    /// Mark the given `UID` as an ancestor of this `Entity`
-    pub fn add_ancestor(&mut self, uid: EntityUID) {
-        self.ancestors.insert(uid);
+    /// Mark the given `UID` as an indirect ancestor of this `Entity`
+    ///
+    /// The given `UID` will not be added as an indirecty ancestor if
+    /// it is already a direct ancestor (parent) of this `Entity`
+    /// The caller of this code is responsible for maintaining
+    /// transitive closure of hierarchy.
+    pub fn add_indirect_ancestor(&mut self, uid: EntityUID) {
+        if !self.parents.contains(&uid) {
+            self.indirect_ancestors.insert(uid);
+        }
     }
 
-    /// Consume the entity and return the entity's owned Uid, attributes, parents, and tags.
+    /// Mark the given `UID` as a (direct) parent of this `Entity`, and
+    /// remove the UID from indirect ancestors
+    /// if it was previously added as an indirect ancestor
+    /// The caller of this code is responsible for maintaining
+    /// transitive closure of hierarchy.
+    pub fn add_parent(&mut self, uid: EntityUID) {
+        self.indirect_ancestors.remove(&uid);
+        self.parents.insert(uid);
+    }
+
+    /// Remove the given `UID` as an indirect ancestor of this `Entity`.
+    ///
+    /// No effect if the `UID` is a direct parent.
+    /// The caller of this code is responsible for maintaining
+    /// transitive closure of hierarchy.
+    pub fn remove_indirect_ancestor(&mut self, uid: &EntityUID) {
+        self.indirect_ancestors.remove(uid);
+    }
+
+    /// Remove the given `UID` as a (direct) parent of this `Entity`.
+    ///
+    /// No effect on the `Entity`'s indirect ancestors.
+    /// The caller of this code is responsible for maintaining
+    /// transitive closure of hierarchy.
+    pub fn remove_parent(&mut self, uid: &EntityUID) {
+        self.parents.remove(uid);
+    }
+
+    /// Consume the entity and return the entity's owned Uid, attributes, ancestors, parents, and tags.
     pub fn into_inner(
         self,
     ) -> (
         EntityUID,
         HashMap<SmolStr, PartialValue>,
         HashSet<EntityUID>,
+        HashSet<EntityUID>,
         HashMap<SmolStr, PartialValue>,
     ) {
         let Self {
             uid,
             attrs,
-            ancestors,
+            indirect_ancestors,
+            parents,
             tags,
         } = self;
         (
             uid,
             attrs.into_iter().map(|(k, v)| (k, v.0)).collect(),
-            ancestors,
+            indirect_ancestors,
+            parents,
             tags.into_iter().map(|(k, v)| (k, v.0)).collect(),
         )
     }
@@ -568,7 +644,7 @@ impl TCNode<EntityUID> for Entity {
     }
 
     fn add_edge_to(&mut self, k: EntityUID) {
-        self.add_ancestor(k)
+        self.add_indirect_ancestor(k);
     }
 
     fn out_edges(&self) -> Box<dyn Iterator<Item = &EntityUID> + '_> {
@@ -587,7 +663,7 @@ impl TCNode<EntityUID> for Arc<Entity> {
 
     fn add_edge_to(&mut self, k: EntityUID) {
         // Use Arc::make_mut to get a mutable reference to the inner value
-        Arc::make_mut(self).add_ancestor(k)
+        Arc::make_mut(self).add_indirect_ancestor(k)
     }
 
     fn out_edges(&self) -> Box<dyn Iterator<Item = &EntityUID> + '_> {
@@ -609,7 +685,7 @@ impl std::fmt::Display for Entity {
                 .iter()
                 .map(|(k, v)| format!("{}: {}", k, v))
                 .join("; "),
-            self.ancestors.iter().join(", ")
+            self.ancestors().join(", ")
         )
     }
 }