+++ cedar-policy-core/src/entities/json/entities.rs
@@ -31,7 +31,10 @@ use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
 use smol_str::SmolStr;
 use std::sync::Arc;
-use std::{collections::HashMap, io::Read};
+use std::{
+    collections::{HashMap, HashSet},
+    io::Read,
+};
 
 #[cfg(feature = "wasm")]
 extern crate tsify;
@@ -416,7 +419,14 @@ impl<'e, 's, S: Schema> EntityJsonParser<'e, 's, S> {
                 })
             })
             .collect::<Result<_, JsonDeserializationError>>()?;
-        Ok(Entity::new(uid, attrs, parents, tags, self.extensions)?)
+        Ok(Entity::new(
+            uid,
+            attrs,
+            HashSet::new(),
+            parents,
+            tags,
+            self.extensions,
+        )?)
     }
 }