+++ cedar-policy-validator/src/schema.rs
@@ -883,6 +883,7 @@ impl ValidatorSchema {
             Entity::new_with_attr_partial_value_serialized_as_expr(
                 action_id.clone(),
                 action.attributes.clone(),
+                HashSet::new(),
                 action_ancestors.remove(action_id).unwrap_or_default(),
                 BTreeMap::new(), // actions cannot have entity tags
             )
@@ -2580,7 +2581,13 @@ pub(crate) mod test {
         let view_photo = actions.entity(&action_uid);
         assert_eq!(
             view_photo.unwrap(),
-            &Entity::new_with_attr_partial_value(action_uid, [], HashSet::new(), [])
+            &Entity::new_with_attr_partial_value(
+                action_uid,
+                [],
+                HashSet::new(),
+                HashSet::new(),
+                []
+            )
         );
     }
 
@@ -2615,26 +2622,16 @@ pub(crate) mod test {
             &Entity::new_with_attr_partial_value(
                 view_photo_uid,
                 [],
+                HashSet::new(),
                 HashSet::from([view_uid.clone(), read_uid.clone()]),
                 [],
-            )
-        );
-
-        let view_entity = actions.entity(&view_uid);
-        assert_eq!(
-            view_entity.unwrap(),
-            &Entity::new_with_attr_partial_value(
-                view_uid,
-                [],
-                HashSet::from([read_uid.clone()]),
-                []
             )
         );
 
         let read_entity = actions.entity(&read_uid);
         assert_eq!(
             read_entity.unwrap(),
-            &Entity::new_with_attr_partial_value(read_uid, [], HashSet::new(), [])
+            &Entity::new_with_attr_partial_value(read_uid, [], HashSet::new(), HashSet::new(), [])
         );
     }
 
@@ -2662,6 +2659,7 @@ pub(crate) mod test {
             &Entity::new(
                 action_uid,
                 [("attr".into(), RestrictedExpr::val("foo"))],
+                HashSet::new(),
                 HashSet::new(),
                 [],
                 Extensions::none(),