+++ cedar-policy/src/api.rs
@@ -131,6 +131,7 @@ impl Entity {
         Self(ast::Entity::new_with_attr_partial_value(
             uid.into(),
             [],
+            [].into_iter().collect(),
             parents.into_iter().map(EntityUid::into).collect(),
             [],
         ))
@@ -151,6 +152,7 @@ impl Entity {
         Ok(Self(ast::Entity::new(
             uid.into(),
             attrs.into_iter().map(|(k, v)| (k.into(), v.0)),
+            [].into_iter().collect(),
             parents.into_iter().map(EntityUid::into).collect(),
             tags.into_iter().map(|(k, v)| (k.into(), v.0)),
             Extensions::all_available(),
@@ -231,7 +233,8 @@ impl Entity {
         HashMap<String, RestrictedExpression>,
         HashSet<EntityUid>,
     ) {
-        let (uid, attrs, ancestors, _) = self.0.into_inner();
+        let (uid, attrs, ancestors, mut parents, _) = self.0.into_inner();
+        parents.extend(ancestors);
 
         let attrs = attrs
             .into_iter()
@@ -253,7 +256,7 @@ impl Entity {
         (
             uid.into(),
             attrs,
-            ancestors.into_iter().map(Into::into).collect(),
+            parents.into_iter().map(Into::into).collect(),
         )
     }
 
@@ -366,7 +369,7 @@ impl Entities {
     }
 
     /// Transform the store into a partial store, where
-    /// attempting to dereference a non-existent `EntityUID` results in
+    /// attempting to dereference a non-existent `EntityUid` results in
     /// a residual instead of an error.
     #[doc = include_str!("../experimental_warning.md")]
     #[must_use]
@@ -443,6 +446,22 @@ impl Entities {
                 Extensions::all_available(),
             )?,
         ))
+    }
+
+    /// Removes each of the [`EntityUid`]s in the iterator
+    /// from this [`Entities`] structure, re-computing the transitive
+    /// closure after removing all edges to/from the removed entities.
+    ///
+    /// Re-computing the transitive closure can be expensive, so it is
+    /// advised to not call this method in a loop.
+    pub fn remove_entities(
+        self,
+        entity_ids: impl IntoIterator<Item = EntityUid>,
+    ) -> Result<Self, EntitiesError> {
+        Ok(Self(self.0.remove_entities(
+            entity_ids.into_iter().map(|euid| euid.0),
+            cedar_policy_core::entities::TCComputation::ComputeNow,
+        )?))
     }
 
     /// Parse an entities JSON file (in [&str] form) and add them into this